backend/requirements.txt
aiohttp==3.8.5
aioredis==2.0.1
ccxt==4.0.77
websockets==11.0.3
pyyaml==6.0.1
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
ta==0.10.2
pytest==7.4.0
pytest-asyncio==0.21.1
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pandas_ta==0.3.14b0
psycopg2-binary==2.9.7
sqlalchemy==2.0.21


